+++ Frends.SFTP.ReadFile/Frends.SFTP.ReadFile.Tests/ConnectivityTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Frends.SFTP.ReadFile.Enums;
 
@@ -9,37 +10,37 @@ namespace Frends.SFTP.ReadFile.Tests;
 public class ConnectivityTests : ReadFileTestBase
 {
     [Test]
-    public void ReadFile_TestWithLargerBuffer()
+    public async Task ReadFile_TestWithLargerBuffer()
     {
         _connection.BufferSize = 256;
 
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_input.Path, result.Path);
         Assert.AreEqual(_content, result.Content);
     }
 
     [Test]
-    public void ReadFile_TestWithPrivateKeyFileRsa()
+    public async Task ReadFile_TestWithPrivateKeyFileRsa()
     {
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_input.Path, result.Path);
         Assert.AreEqual(_content, result.Content);
     }
 
     [Test]
-    public void ReadFile_TestWithPrivateKeyFileRsaFromString()
+    public async Task ReadFile_TestWithPrivateKeyFileRsaFromString()
     {
         var key = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
 
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyString = key;
 
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_input.Path, result.Path);
         Assert.AreEqual(_content, result.Content);
     }