+++ Frends.SFTP.ReadFile/Frends.SFTP.ReadFile.Tests/ErrorTests.cs
@@ -21,7 +21,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<SftpPathNotFoundException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<SftpPathNotFoundException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.AreEqual($"No such file", ex.Message);
 
     }
@@ -37,7 +37,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        Assert.Throws<SocketException>(() => SFTP.ReadFile(input, connection));
+        Assert.ThrowsAsync<SocketException>(async () => await SFTP.ReadFile(input, connection, default));
     }
 
     [Test]
@@ -53,7 +53,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<SshAuthenticationException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<SshAuthenticationException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.AreEqual("Permission denied (password).", ex.Message);
     }
 
@@ -62,7 +62,7 @@ public class ErrorTests
     {
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        connection.PrivateKeyFilePassphrase = "demo";
+        connection.PrivateKeyPassphrase = "demo";
         connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
         var input = new Input
@@ -71,7 +71,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info:"));
     }
 
@@ -80,7 +80,7 @@ public class ErrorTests
     {
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyFile = "";
 
         var input = new Input
@@ -89,7 +89,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 
@@ -100,7 +100,7 @@ public class ErrorTests
 
         var connection = Helpers.GetSftpConnection();
         connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        connection.PrivateKeyFilePassphrase = "passphrase";
+        connection.PrivateKeyPassphrase = "passphrase";
         connection.PrivateKeyString = key.ToString();
 
         var input = new Input
@@ -109,7 +109,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.That(ex.Message.StartsWith("Error when initializing connection info: "));
     }
 
@@ -127,7 +127,7 @@ public class ErrorTests
             FileEncoding = FileEncoding.ANSI,
         };
 
-        var ex = Assert.Throws<SshConnectionException>(() => SFTP.ReadFile(input, connection));
+        var ex = Assert.ThrowsAsync<SshConnectionException>(async () => await SFTP.ReadFile(input, connection, default));
         Assert.AreEqual("Key exchange negotiation failed.", ex.Message);
     }
 }