+++ Frends.SFTP.ReadFile/Frends.SFTP.ReadFile.Tests/ReadTests.cs
@@ -1,3 +1,4 @@
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Frends.SFTP.ReadFile.Enums;
 
@@ -7,54 +8,54 @@ namespace Frends.SFTP.ReadFile.Tests;
 class ReadTests : ReadFileTestBase
 {
     [Test]
-    public void ReadFile_TestSimpleRead()
+    public async Task ReadFile_TestSimpleRead()
     {
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
     }
 
     [Test]
-    public void ReadFile_TestWithEmptyContent()
+    public async Task ReadFile_TestWithEmptyContent()
     {
         var content = "";
 
         Helpers.OverrideDummyFile(_input.Path, content);
 
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(string.Empty, result.Content);
     }
 
     [Test]
-    public void ReadFile_TestWithDifferentEncoding()
+    public async Task ReadFile_TestWithDifferentEncoding()
     {
         _input.FileEncoding = FileEncoding.ANSI;
 
         Helpers.GenerateDummyFile(_input.Path, _content);
 
-        var result = SFTP.ReadFile(_input, _connection);
+        var result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
         
         _input.FileEncoding = FileEncoding.ASCII;
-        result = SFTP.ReadFile(_input, _connection);
+        result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
 
         _input.FileEncoding = FileEncoding.UTF8;
         _input.EnableBom = true;
-        result = SFTP.ReadFile(_input, _connection);
+        result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
 
         _input.FileEncoding = FileEncoding.UTF8;
         _input.EnableBom = false;
-        result = SFTP.ReadFile(_input, _connection);
+        result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
 
         _input.FileEncoding = FileEncoding.WINDOWS1252;
-        result = SFTP.ReadFile(_input, _connection);
+        result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
 
         _input.FileEncoding = FileEncoding.Other;
         _input.EncodingInString = "iso-8859-1";
-        result = SFTP.ReadFile(_input, _connection);
+        result = await SFTP.ReadFile(_input, _connection, default);
         Assert.AreEqual(_content, result.Content);
     }
 }